i101: Extract notifications logic to hook
