+++ react-app/src/library/hooks/index.js
@@ -1,7 +1,8 @@
 export { default as useLocation } from './useLocation';
 export { default as useMap } from './useMap';
-export { default as useApp } from './useApp';
+export { default as useAppLaunch } from './useAppLaunch';
 export { default as useAnimation } from './useAnimation';
 export { default as useNewShout } from './useNewShout';
+export { default as useNotifications } from './useNotifications';
 export { default as useSettings } from './useSettings';
 export { default as useShouts } from './useShouts';
